make go-gettable
