+++ .gitignore
@@ -6,3 +6,4 @@ files/*
 src/github.com
 # Binary
 /wiki
+/g-wiki
